@@ -55,65 +55,67 @@
     "release": "standard-version"
   },
   "dependencies": {
-    "@angular/animations": "^5.2.2",
-    "@angular/cdk": "^5.1.0",
-    "@angular/common": "^5.2.2",
-    "@angular/compiler": "^5.2.2",
-    "@angular/core": "^5.2.2",
-    "@angular/forms": "^5.2.2",
-    "@angular/http": "^5.2.2",
-    "@angular/material": "^5.1.0",
-    "@angular/platform-browser": "^5.2.2",
-    "@angular/platform-browser-dynamic": "^5.2.2",
-    "@angular/platform-server": "^5.2.2",
-    "@angular/router": "^5.2.2",
-    "@angular/service-worker": "^5.2.2",
-    "@ngrx/store": "^4.1.1",
-    "@nguniversal/express-engine": "^5.0.0-beta.5",
-    "@nguniversal/module-map-ngfactory-loader": "^5.0.0-beta.5",
-    "core-js": "^2.4.1",
-    "express": "^4.16.2",
-    "rxjs": "^5.5.2",
-    "ts-loader": "^3.2.0",
-    "zone.js": "^0.8.14"
+    "@angular/animations": "^6.0.3",
+    "@angular/cdk": "^6.2.0",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/material": "^6.2.0",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/platform-server": "^6.0.3",
+    "@angular/router": "^6.0.3",
+    "@angular/service-worker": "^6.0.3",
+    "@ngrx/store": "^6.0.0",
+    "@nguniversal/express-engine": "^6.0.0",
+    "@nguniversal/module-map-ngfactory-loader": "^6.0.0",
+    "core-js": "^2.5.7",
+    "express": "^4.16.3",
+    "rxjs": "^6.2.0",
+    "ts-loader": "^4.3.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.6",
-    "@angular/compiler-cli": "^5.2.2",
-    "@angular/language-service": "^5.2.2",
-    "@ngrx/store-devtools": "^4.1.1",
-    "@types/jasmine": "~2.5.53",
+    "@angular-devkit/build-angular": "~0.6.6",
+    "@angular/cli": "^6.0.7",
+    "@angular/compiler-cli": "^6.0.3",
+    "@angular/language-service": "^6.0.3",
+    "@ngrx/store-devtools": "^6.0.1",
+    "@types/express": "^4.11.1",
+    "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "coveralls": "^3.0.0",
+    "@types/node": "~10.1.2",
+    "codelyzer": "^4.3.0",
+    "coveralls": "^3.0.1",
     "cz-conventional-changelog": "^2.1.0",
-    "electron": "^1.7.11",
-    "electron-packager": "^10.1.1",
+    "electron": "^2.0.2",
+    "electron-packager": "^12.1.0",
     "fs-readdir-recursive": "^1.1.0",
     "husky": "^0.14.3",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "jasmine-core": "~3.1.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.2",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "^2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^1.1.0",
     "karma-spec-reporter": "0.0.32",
     "mkdir-recursive": "^0.4.0",
-    "ngrx-store-freeze": "^0.2.0",
-    "prettier": "^1.10.2",
-    "pretty-quick": "^1.2.2",
-    "protractor": "~5.1.2",
-    "standard-version": "^4.3.0",
-    "stylelint": "^8.4.0",
-    "stylelint-config-recommended-scss": "^3.0.0",
-    "stylelint-config-standard": "^18.0.0",
-    "stylelint-scss": "^2.2.0",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.9.1",
-    "typescript": "^2.6.2"
+    "ngrx-store-freeze": "^0.2.3",
+    "prettier": "^1.13.4",
+    "pretty-quick": "^1.6.0",
+    "protractor": "~5.3.2",
+    "standard-version": "^4.4.0",
+    "stylelint": "^9.2.1",
+    "stylelint-config-recommended-scss": "^3.2.0",
+    "stylelint-config-standard": "^18.2.0",
+    "stylelint-scss": "^3.1.0",
+    "ts-node": "~6.0.5",
+    "tslint": "~5.10.0",
+    "typescript": "~2.7.2"
   },
   "config": {
     "commitizen": {
