@@ -1,8 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { NO_ERRORS_SCHEMA, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable, of } from 'rxjs';
 import { Store } from '@ngrx/store';
-import 'rxjs/add/observable/of';
 
 import { TodoComponent } from './todo.component';
 import { Todo } from '../shared/todo.model';
@@ -15,22 +14,20 @@ describe('TodoComponent', () => {
   let component: TodoComponent;
   let fixture: ComponentFixture<TodoComponent>;
 
-  beforeEach(
-    async(() => {
-      mockStore = {
-        select: jasmine.createSpy('select').and.returnValue(Observable.of(0)),
-        dispatch: jasmine.createSpy('dispatch')
-      };
+  beforeEach(async(() => {
+    mockStore = {
+      select: jasmine.createSpy('select').and.returnValue(of(0)),
+      dispatch: jasmine.createSpy('dispatch')
+    };
 
-      todo = { id: 0, text: 'foo', completed: false };
+    todo = { id: 0, text: 'foo', completed: false };
 
-      TestBed.configureTestingModule({
-        providers: [{ provide: Store, useValue: mockStore }],
-        schemas: [NO_ERRORS_SCHEMA, CUSTOM_ELEMENTS_SCHEMA],
-        declarations: [TodoComponent]
-      }).compileComponents();
-    })
-  );
+    TestBed.configureTestingModule({
+      providers: [{ provide: Store, useValue: mockStore }],
+      schemas: [NO_ERRORS_SCHEMA, CUSTOM_ELEMENTS_SCHEMA],
+      declarations: [TodoComponent]
+    }).compileComponents();
+  }));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(TodoComponent);
